moved db operations to a new service and updated pom
