+++ DatabaseProxyService/src/main/java/com/DatabaseProxyService/DatabaseProxyServiceApplication.java
@@ -1,4 +1,4 @@
-package com.DatabaseProxyService;
+package com.databaseproxyservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
