+++ DatabaseProxyService/src/main/java/com/DatabaseProxyService/service/ConnectionPoolService.java
@@ -1,5 +1,8 @@
 package com.databaseproxyservice.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -10,6 +13,7 @@ public class ConnectionPoolService {
 
     private final Map<String, DataSource> dataSources;
 
+    @Autowired
     public ConnectionPoolService(Map<String, DataSource> dataSources) {
         this.dataSources = dataSources;
     }
