+++ DatabaseProxyService/src/main/resources/application.yml
@@ -1,14 +1,12 @@
 spring:
-  application:
-    name: DatabaseProxyService
   databases:
     db1:
-      url: jdbc:mysql://localhost:3306/db1
+      url: jdbc:oracle:thin:@//localhost:3306/db1
       username: user1
       password: password1
-      driver-class-name: com.mysql.cj.jdbc.Driver
+      driver-class-name: oracle.jdbc.OracleDriver
     db2:
-      url: jdbc:postgresql://localhost:5432/db2
+      url: jdbc:oracle:thin:@//localhost:5432/db2
       username: user2
       password: password2
-      driver-class-name: org.postgresql.Driver
+      driver-class-name: oracle.jdbc.OracleDriver
