+++ DatabaseProxyService/src/test/java/com/DatabaseProxyService/DatabaseProxyServiceApplicationTests.java
@@ -1,13 +1,14 @@
 package com.DatabaseProxyService;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class DatabaseProxyServiceApplicationTests {
+@SpringBootTest()
+public class DatabaseProxyServiceApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+//	@Test
+//	void contextLoads() {
+//	}
 
 }
